+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### 1.1.0 (03-04-2024)
+
+- Add Carousel Component
+- Exchange lucide-Icons with FontAwesomeIcons
+
 ### 0.0.7 (15-03-2024)
 
 - Handle Reference by forwardRef in Select component